refactor(connector): re-arrange the `connector::parser` module for better readibility (#19697)
