+++ src/config/docs.md
@@ -117,7 +117,7 @@ This page is automatically generated by `./risedev generate-example-config`
 | cache_refill |  |  |
 | check_compaction_result |  | false |
 | compact_iter_recreate_timeout_ms |  | 600000 |
-| compactor_concurrent_uploading_sst_count | The concurrent uploading number of `SSTables` of buidler |  |
+| compactor_concurrent_uploading_sst_count | The concurrent uploading number of `SSTables` of builder |  |
 | compactor_fast_max_compact_delete_ratio |  | 40 |
 | compactor_fast_max_compact_task_size |  | 2147483648 |
 | compactor_iter_max_io_retry_times |  | 8 |
